Fixed some types to size_t

+++ src/EntityType.cpp
@@ -35,7 +35,7 @@ EntityType::EntityType(SDL_RWops *stream)
 
 void EntityType::init(void)
 {
-	int varSize;
+	size_t varSize;
 	Uint32 *startData;
 	getVars(&varSize, &startData);	
 	memset(startData,0,varSize*sizeof(Uint32));
@@ -43,10 +43,10 @@ void EntityType::init(void)
 
 void EntityType::load(SDL_RWops *stream)
 {
-	int size;
+	size_t size;
 	Uint32 *startData;
 	getVars(&size, &startData);
-	for (int i=0;i<size;i++)
+	for (size_t i=0;i<size;i++)
 		startData[i]=SDL_ReadBE32(stream);
 }
 
@@ -57,7 +57,7 @@ bool EntityType::loadText(SDL_RWops *stream)
 	char *varname;
 	int val;
 
-	int varSize;
+	size_t varSize;
 	Uint32 *startData;
 	const char **tab=getVars(&varSize, &startData);	
 
@@ -78,7 +78,7 @@ bool EntityType::loadText(SDL_RWops *stream)
 		else
 			val=0;
 
-		for (int i=0; i<varSize; i++)
+		for (size_t i=0; i<varSize; i++)
 			if (strcmp(tab[i],varname)==0)
 			{
 				*(startData+i)=val;
@@ -89,21 +89,20 @@ bool EntityType::loadText(SDL_RWops *stream)
 
 void EntityType::save(SDL_RWops *stream)
 {
-	int size;
+	size_t size;
 	Uint32 *startData;
 	getVars(&size, &startData);
-	for (int i=0;i<size;i++)
+	for (size_t i=0;i<size;i++)
 		SDL_WriteBE32(stream, startData[i]);
 }
 
 void EntityType::dump(void)
 {
-
-	int varSize;
+	size_t varSize;
 	Uint32 *startData;
 	const char **tab=getVars(&varSize, &startData);
 
 	printf("%d Elements :\n",varSize);
-	for (int i=0; i<varSize;i++)
+	for (size_t i=0; i<varSize;i++)
 		printf("\t%s = %d\n",tab[i],*(startData+i));
 }
